fix: refactor folder structure
