+++ src/api/assets/mod.rs
