+++ src/api/auth/signout.rs
