+++ src/api/dashboard/user_settings.rs
@@ -8,22 +8,22 @@ use reqwest::StatusCode;
 
 use crate::{
     domain::{AppState, Token},
-    localization::Translator,
+    util::localization::Translator,
 };
 
 #[derive(Template)]
-#[template(path = "dashboard/groceries/index.html")]
-struct GroceriesTemplate {
+#[template(path = "dashboard/settings.html")]
+struct SettingsTemplate {
     translator: Translator,
     current_year: i32,
     email: String,
 }
 
-pub async fn groceries_page_handler(
+pub async fn settings_page_handler(
     State(state): State<AppState>,
     Token(claims): Token,
 ) -> impl IntoResponse {
-    let template = GroceriesTemplate {
+    let template = SettingsTemplate {
         translator: state.translator.clone(),
         current_year: chrono::Utc::now().year(),
         email: claims.email,
