+++ src/api/mod.rs
@@ -3,3 +3,4 @@ pub mod auth;
 pub mod dashboard;
 pub mod generic;
 pub mod home;
+pub mod layer;
