+++ src/components/auth/login_failed_message.rs
@@ -3,5 +3,5 @@ use askama::Template;
 #[derive(Template)]
 #[template(path = "auth/login/login_failed_message.html")]
 pub struct LoginFailedMessageTemplate {
-    pub translator: crate::localization::Translator,
+    pub translator: crate::util::localization::Translator,
 }
